+++ CHANGELOG.md
@@ -7,3 +7,4 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - New method ``expression.Indices.__getitem__()``.
+- New method ``loperator.hilbert_space.index()``.